server/requirements.txt
beautifulsoup4==4.12.2
blinker==1.7.0
cachelib==0.9.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
Flask==3.0.0
Flask-Caching==2.1.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
soupsieve==2.5
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.16.0
